feat(mnemosyne): Randomize free-list reuse and sweep thread-count scaling
Rescued from an uncommitted working tree two hours past its author's last
board update. Sixty-two files, verified here and completed to the gate
rather than extended.

The allocator gains a secondary free list per page and a single decision
point, `Page::choose_free_head`, that selects between the two. Selection
randomizes whenever the secondary list is non-empty, whatever the
allocating policy asks for, so a freed block's address stops being
predictable to a caller that frees and immediately reallocates. Reuse
remains reachable to every policy: the owning segment's recorded mode, not
the caller's, still decides how a link is encoded (ADR 0001).

The benchmark suite gains a `Thread count scaling` group sweeping 1, 2, 4,
8 and 16 workers through the same cycle path, and
`benchmarks/allocator_comparison.md` opens with the gaps the table makes
visible -- huge-object teardown at 4477 ns against MiMalloc's 104 ns,
scaling shape hidden behind a single fixed-thread row, and no mixed-size
churn workload at all.

Completed to land:

- Two `policy_integration_tests` cases were red. Both asserted pointer
  identity on reuse as the proxy for "the link remains decodable", which
  randomization invalidates without touching the property. They now assert
  the property: a run of allocations off the affected chain comes back
  distinct, aligned, disjoint from every still-live block, and each
  round-trips a payload written through it -- a mis-decoded link hands out
  a wild or overlapping block. The owner-block realloc case asserts payload
  survival across a size-class move instead of an unchanged address.
  ADR 0001 carries a dated revision note recording that its verification,
  not its decision, is what changed.
- Seven new `unsafe fn` carried no `# Safety` section; each now states what
  the caller must guarantee, and for the cookie-taking pair, that a wrong
  cookie is undetectable at the call and corrupts the chain at first use.
- Twenty-eight duplicated generic bounds fold into their where clauses,
  three alignment checks take `is_multiple_of`, one `let...else` becomes
  `?`, and three test `unwrap`s state their invariant.
- `AsmTls::register_current_allocator_ptr` carries an `expect` with its
  reason: `set_teb_tls_slot` writes the pointer value into a TEB slot and
  never reads through it.

Verified at this revision: 459/459 workspace tests, `cargo clippy
--workspace --all-targets -- -D warnings` clean, `cargo fmt --check` clean.

Item: MN-RANDOMIZED-FREE-LIST-GUARD-REGRESSION
Refs: backlog.md#mn-randomized-free-list-guard-regression

+++ crates/mnemosyne-arena/src/segment/pool/huge_pool.rs
@@ -128,7 +128,7 @@ impl GlobalHugePool {
         // SAFETY: by this function's contract `segment` is a valid, initialized,
         // exclusively-owned huge-allocation `Segment`, so reading its page-0
         // `block_size` is sound.
-        let size = unsafe { (*segment).pages[0].block_size };
+        let size = unsafe { (*segment).pages[0].block_size as usize };
         if size > Self::MAX_CACHED_HUGE_SIZE {
             return false;
         }
@@ -285,7 +285,7 @@ impl GlobalHugePool {
             if let Some(segment) = popped {
                 // SAFETY: the pop transferred exclusive ownership of `segment`
                 // to this caller, so reading its page-0 `block_size` is sound.
-                let block_size = unsafe { (*segment).pages[0].block_size };
+                let block_size = unsafe { (*segment).pages[0].block_size as usize };
                 pool_node
                     .total_count
                     .value
@@ -325,7 +325,7 @@ impl GlobalHugePool {
         let mut fit = None;
         while let Some(segment) = bucket.pop_head(band) {
             // SAFETY: `pop_head` transfers exclusive ownership of `segment`.
-            let block_size = unsafe { (*segment).pages[0].block_size };
+            let block_size = unsafe { (*segment).pages[0].block_size as usize };
             if block_size >= size {
                 fit = Some(segment);
                 break;
@@ -444,7 +444,7 @@ impl GlobalHugePool {
                                 .next_free_segment
                                 .load(core::sync::atomic::Ordering::Relaxed);
                             let raw_ptr = (*head).raw_alloc_ptr;
-                            let block_size = (*head).pages[0].block_size;
+                            let block_size = (*head).pages[0].block_size as usize;
                             let _ = B::deallocate(raw_ptr, block_size);
                             next
                         };
